add TF Mean, Min, Sum, Prod
